move Receptor into receptors directory
